+++ Petlil/src/main/java/br/lil/dao/ServicoDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.lil.model.Servico;
 
-public interface ServicoDao extends JpaRepository<Servico, Long> {
+public interface ServicoDao extends JpaRepository<Servico, Integer> {
     
 }
